+++ mzLib/Test/Omics/Modifications/TestModFits.cs
@@ -5,7 +5,7 @@ using System;
 using Omics.Modifications;
 using Stopwatch = System.Diagnostics.Stopwatch;
 
-namespace Test
+namespace Test.Omics.Modifications
 {
     [TestFixture]
     [System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage]
